@@ -10,10 +10,10 @@ readme = "README.md"
 authors = [{ name = "Fabian Hofmann", email = "fabianmarikhofmann@gmail.com" }]
 license = { file = "LICENSE" }
 classifiers = [
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Development Status :: 3 - Alpha",
     "Environment :: Console",
     "Intended Audience :: Science/Research",
@@ -23,7 +23,7 @@ classifiers = [
     "Operating System :: OS Independent",
 ]
 
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
     "numpy; python_version > '3.10'",
     "numpy<2; python_version <= '3.10'",
@@ -36,6 +36,8 @@ dependencies = [
     "polars",
     "tqdm",
     "deprecation",
+    "google-cloud-storage",
+    "requests"
 ]
 
 [project.urls]
@@ -51,6 +53,8 @@ docs = [
     "sphinx_book_theme==1.1.3",
     "nbsphinx==0.9.4",
     "nbsphinx-link==1.3.0",
+    "docutils<0.21",
+    "numpy<2",
     "gurobipy==11.0.2",
     "ipykernel==6.29.5",
     "matplotlib==3.9.1",
@@ -63,6 +67,7 @@ dev = [
     "netcdf4",
     "paramiko",
     "types-paramiko",
+    "types-requests",
     "gurobipy",
     "highspy",
 ]
@@ -87,6 +92,10 @@ include = ["linopy"]
 [tool.setuptools_scm]
 write_to = "linopy/version.py"
 version_scheme = "no-guess-dev"
+
+[tool.pytest.ini_options]
+testpaths = ["test"]
+norecursedirs = ["dev-scripts", "doc", "examples", "benchmark"]
 
 [tool.coverage.run]
 branch = true
