@@ -1,21 +1,14 @@
 from __future__ import annotations
 
-import sys
 from collections.abc import Hashable, Iterable, Mapping, Sequence
 from pathlib import Path
 from typing import TYPE_CHECKING, Union
 
 import numpy
-import numpy.typing
+import polars as pl
 from pandas import DataFrame, Index, Series
 from xarray import DataArray
 from xarray.core.coordinates import DataArrayCoordinates, DatasetCoordinates
-
-if sys.version_info >= (3, 10):
-    from types import EllipsisType, NotImplementedType
-else:
-    EllipsisType = type(Ellipsis)
-    NotImplementedType = type(NotImplemented)
 
 if TYPE_CHECKING:
     from linopy.constraints import AnonymousScalarConstraint, Constraint
@@ -27,15 +20,15 @@ if TYPE_CHECKING:
     from linopy.variables import ScalarVariable, Variable
 
 # Type aliases using Union for Python 3.9 compatibility
-CoordsLike = Union[
-    Sequence[Union[Sequence, Index, DataArray]],
+CoordsLike = Union[  # noqa: UP007
+    Sequence[Sequence | Index | DataArray],
     Mapping,
     DataArrayCoordinates,
     DatasetCoordinates,
 ]
-DimsLike = Union[str, Iterable[Hashable]]
+DimsLike = Union[str, Iterable[Hashable]]  # noqa: UP007
 
-ConstantLike = Union[
+ConstantLike = Union[  # noqa: UP007
     int,
     float,
     numpy.floating,
@@ -44,8 +37,9 @@ ConstantLike = Union[
     DataArray,
     Series,
     DataFrame,
+    pl.Series,
 ]
-SignLike = Union[str, numpy.ndarray, DataArray, Series, DataFrame]
+SignLike = Union[str, numpy.ndarray, DataArray, Series, DataFrame]  # noqa: UP007
 VariableLike = Union["ScalarVariable", "Variable"]
 ExpressionLike = Union[
     "ScalarLinearExpression",
@@ -53,6 +47,6 @@ ExpressionLike = Union[
     "QuadraticExpression",
 ]
 ConstraintLike = Union["Constraint", "AnonymousScalarConstraint"]
-MaskLike = Union[numpy.ndarray, DataArray, Series, DataFrame]
-SideLike = Union[ConstantLike, VariableLike, ExpressionLike]
-PathLike = Union[str, Path]
+MaskLike = Union[numpy.ndarray, DataArray, Series, DataFrame]  # noqa: UP007
+SideLike = Union[ConstantLike, VariableLike, ExpressionLike]  # noqa: UP007
+PathLike = Union[str, Path]  # noqa: UP007
