@@ -9,13 +9,11 @@ from __future__ import annotations
 
 import functools
 import logging
+from abc import ABC, abstractmethod
 from collections.abc import Callable, Hashable, Iterator, Mapping, Sequence
 from dataclasses import dataclass, field
 from itertools import product, zip_longest
-from typing import (
-    TYPE_CHECKING,
-    Any,
-)
+from typing import TYPE_CHECKING, Any, TypeVar, overload
 from warnings import warn
 
 import numpy as np
@@ -40,6 +38,8 @@ try:
 except ImportError:
     import xarray.core.rolling
     from xarray.core.rolling import DatasetRolling  # type: ignore
+
+from types import EllipsisType, NotImplementedType
 
 from linopy import constraints, variables
 from linopy.common import (
@@ -79,9 +79,7 @@ from linopy.constants import (
 from linopy.types import (
     ConstantLike,
     DimsLike,
-    EllipsisType,
     ExpressionLike,
-    NotImplementedType,
     SideLike,
     SignLike,
     VariableLike,
@@ -100,6 +98,7 @@ SUPPORTED_CONSTANT_TYPES = (
     pd.Series,
     pd.DataFrame,
     np.ndarray,
+    pl.Series,
 )
 
 
@@ -131,7 +130,7 @@ def exprwrap(
 def _expr_unwrap(
     maybe_expr: Any | LinearExpression | QuadraticExpression,
 ) -> Any:
-    if isinstance(maybe_expr, (LinearExpression, QuadraticExpression)):
+    if isinstance(maybe_expr, LinearExpression | QuadraticExpression):
         return maybe_expr.data
 
     return maybe_expr
@@ -251,6 +250,8 @@ class LinearExpressionGroupby:
                 orig_group = group
                 group = group.apply(tuple, axis=1).map(int_map)
 
+            # At this point, group is always a pandas Series
+            assert isinstance(group, pd.Series)
             group_dim = group.index.name
 
             arrays = [group, group.groupby(group).cumcount()]
@@ -317,44 +318,7 @@ class LinearExpressionRolling:
         return LinearExpression(ds, self.model)
 
 
-class LinearExpression:
-    """
-    A linear expression consisting of terms of coefficients and variables.
-
-    The LinearExpression class is a subclass of xarray.Dataset which allows to
-    apply most xarray functions on it. However most arithmetic operations are
-    overwritten. Like this you can easily expand and modify the linear
-    expression.
-
-    Examples
-    --------
-    >>> from linopy import Model
-    >>> import pandas as pd
-    >>> m = Model()
-    >>> x = m.add_variables(pd.Series([0, 0]), 1, name="x")
-    >>> y = m.add_variables(4, pd.Series([8, 10]), name="y")
-
-    Combining expressions:
-
-    >>> expr = 3 * x
-    >>> type(expr)
-    <class 'linopy.expressions.LinearExpression'>
-
-    >>> other = 4 * y
-    >>> type(expr + other)
-    <class 'linopy.expressions.LinearExpression'>
-
-    Multiplying:
-
-    >>> type(3 * expr)
-    <class 'linopy.expressions.LinearExpression'>
-
-    Summation over dimensions
-
-    >>> type(expr.sum(dim="dim_0"))
-    <class 'linopy.expressions.LinearExpression'>
-    """
-
+class BaseExpression(ABC):
     __slots__ = ("_data", "_model")
     __array_ufunc__ = None
     __array_priority__ = 10000
@@ -362,6 +326,13 @@ class LinearExpression:
 
     _fill_value = FILL_VALUE
     _data: Dataset
+
+    @property
+    @abstractmethod
+    def flat(self) -> pd.DataFrame: ...
+
+    @abstractmethod
+    def to_polars(self) -> pl.DataFrame: ...
 
     def __init__(self, data: Dataset | Any | None, model: Model) -> None:
         from linopy.model import Model
@@ -487,71 +458,45 @@ class LinearExpression:
             )
             print(self)
 
-    def __add__(self, other: SideLike) -> LinearExpression:
-        """
-        Add an expression to others.
+    @abstractmethod
+    def __add__(
+        self: GenericExpression, other: SideLike
+    ) -> GenericExpression | QuadraticExpression: ...
 
-        Note: If other is a numpy array or pandas object without axes names,
-        dimension names of self will be filled in other
-        """
-        try:
-            if np.isscalar(other):
-                return self.assign(const=self.const + other)
+    @abstractmethod
+    def __radd__(self: GenericExpression, other: SideLike) -> GenericExpression: ...
 
-            other = as_expression(other, model=self.model, dims=self.coord_dims)
-            return merge([self, other], cls=self.__class__)
-        except TypeError:
-            return NotImplemented
+    @abstractmethod
+    def __sub__(
+        self: GenericExpression, other: SideLike
+    ) -> GenericExpression | QuadraticExpression: ...
 
-    def __radd__(self, other: int) -> LinearExpression | NotImplementedType:
-        # This is needed for using python's sum function
-        return self if other == 0 else NotImplemented
+    @abstractmethod
+    def __rsub__(self: GenericExpression, other: SideLike) -> GenericExpression: ...
 
-    def __sub__(self, other: SideLike) -> LinearExpression:
-        """
-        Subtract others from expression.
+    @abstractmethod
+    def __mul__(
+        self: GenericExpression, other: SideLike
+    ) -> GenericExpression | QuadraticExpression: ...
 
-        Note: If other is a numpy array or pandas object without axes names,
-        dimension names of self will be filled in other
-        """
-        try:
-            if np.isscalar(other):
-                return self.assign_multiindex_safe(const=self.const - other)
+    @abstractmethod
+    def __rmul__(
+        self: GenericExpression, other: SideLike
+    ) -> GenericExpression | QuadraticExpression: ...
 
-            other = as_expression(other, model=self.model, dims=self.coord_dims)
-            return merge([self, -other], cls=self.__class__)
-        except TypeError:
-            return NotImplemented
+    @abstractmethod
+    def __matmul__(
+        self: GenericExpression, other: SideLike
+    ) -> GenericExpression | QuadraticExpression: ...
 
-    def __neg__(self) -> LinearExpression | QuadraticExpression:
+    @abstractmethod
+    def __pow__(self, other: int) -> QuadraticExpression: ...
+
+    def __neg__(self: GenericExpression) -> GenericExpression:
         """
         Get the negative of the expression.
         """
         return self.assign_multiindex_safe(coeffs=-self.coeffs, const=-self.const)
-
-    def __mul__(
-        self,
-        other: SideLike,
-    ) -> LinearExpression | QuadraticExpression:
-        """
-        Multiply the expr by a factor.
-        """
-        try:
-            if isinstance(other, QuadraticExpression):
-                raise TypeError(
-                    "unsupported operand type(s) for *: "
-                    f"{type(self)} and {type(other)}. "
-                    "Higher order non-linear expressions are not yet supported."
-                )
-            elif isinstance(other, (variables.Variable, variables.ScalarVariable)):
-                other = other.to_linexpr()
-
-            if isinstance(other, (LinearExpression, ScalarLinearExpression)):
-                return self._multiply_by_linear_expression(other)
-            else:
-                return self._multiply_by_constant(other)
-        except TypeError:
-            return NotImplemented
 
     def _multiply_by_linear_expression(
         self, other: LinearExpression | ScalarLinearExpression
@@ -566,66 +511,43 @@ class LinearExpression:
         # merge on factor dimension only returns v1 * v2 + c1 * c2
         ds = other.data[["coeffs", "vars"]].sel(_term=0).broadcast_like(self.data)
         ds = assign_multiindex_safe(ds, const=other.const)
-        res = merge([self, ds], dim=FACTOR_DIM, cls=QuadraticExpression)
+        res = merge([self, ds], dim=FACTOR_DIM, cls=QuadraticExpression)  # type: ignore
         # deal with cross terms c1 * v2 + c2 * v1
         if self.has_constant:
             res = res + self.const * other.reset_const()
         if other.has_constant:
             res = res + self.reset_const() * other.const
-        return res  # type: ignore
+        return res
 
-    def _multiply_by_constant(self, other: ConstantLike) -> LinearExpression:
+    def _multiply_by_constant(
+        self: GenericExpression, other: ConstantLike
+    ) -> GenericExpression:
         multiplier = as_dataarray(other, coords=self.coords, dims=self.coord_dims)
         coeffs = self.coeffs * multiplier
         assert all(coeffs.sizes[d] == s for d, s in self.coeffs.sizes.items())
         const = self.const * multiplier
         return self.assign(coeffs=coeffs, const=const)
 
-    def __pow__(self, other: int) -> QuadraticExpression:
-        """
-        Power of the expression with a coefficient. The only coefficient allowed is 2.
-        """
-        if not other == 2:
-            raise ValueError("Power must be 2.")
-        return self * self  # type: ignore
-
-    def __rmul__(self, other: ConstantLike) -> LinearExpression | QuadraticExpression:
-        """
-        Right-multiply the expr by a factor.
-        """
-        return self.__mul__(other)
-
-    def __matmul__(
-        self, other: LinearExpression | Variable | ndarray | DataArray
-    ) -> LinearExpression | QuadraticExpression:
-        """
-        Matrix multiplication with other, similar to xarray dot.
-        """
-        if not isinstance(other, (LinearExpression, variables.Variable)):
-            other = as_dataarray(other, coords=self.coords, dims=self.coord_dims)
-
-        common_dims = list(set(self.coord_dims).intersection(other.dims))
-        return (self * other).sum(dim=common_dims)
-
-    def __div__(
-        self, other: Variable | ConstantLike
-    ) -> LinearExpression | QuadraticExpression:
+    def __div__(self: GenericExpression, other: SideLike) -> GenericExpression:
         try:
             if isinstance(
-                other, (LinearExpression, variables.Variable, variables.ScalarVariable)
+                other,
+                variables.Variable
+                | variables.ScalarVariable
+                | LinearExpression
+                | ScalarLinearExpression
+                | QuadraticExpression,
             ):
                 raise TypeError(
                     "unsupported operand type(s) for /: "
                     f"{type(self)} and {type(other)}"
                     "Non-linear expressions are not yet supported."
                 )
-            return self.__mul__(1 / other)
+            return self._multiply_by_constant(other=1 / other)
         except TypeError:
             return NotImplemented
 
-    def __truediv__(
-        self, other: Variable | ConstantLike
-    ) -> LinearExpression | QuadraticExpression:
+    def __truediv__(self: GenericExpression, other: SideLike) -> GenericExpression:
         return self.__div__(other)
 
     def __le__(self, rhs: SideLike) -> Constraint:
@@ -647,27 +569,33 @@ class LinearExpression:
             "Inequalities only ever defined for >= rather than >."
         )
 
-    def add(self, other: SideLike) -> LinearExpression:
+    def add(
+        self: GenericExpression, other: SideLike
+    ) -> GenericExpression | QuadraticExpression:
         """
         Add an expression to others.
         """
         return self.__add__(other)
 
-    def sub(self, other: SideLike) -> LinearExpression:
+    def sub(
+        self: GenericExpression, other: SideLike
+    ) -> GenericExpression | QuadraticExpression:
         """
         Subtract others from expression.
         """
         return self.__sub__(other)
 
-    def mul(self, other: SideLike) -> LinearExpression | QuadraticExpression:
+    def mul(
+        self: GenericExpression, other: SideLike
+    ) -> GenericExpression | QuadraticExpression:
         """
         Multiply the expr by a factor.
         """
         return self.__mul__(other)
 
     def div(
-        self, other: Variable | float | int
-    ) -> LinearExpression | QuadraticExpression:
+        self: GenericExpression, other: VariableLike | ConstantLike
+    ) -> GenericExpression | QuadraticExpression:
         """
         Divide the expr by a factor.
         """
@@ -679,15 +607,17 @@ class LinearExpression:
         """
         return self.__pow__(other)
 
-    def dot(self, other: ndarray) -> LinearExpression:
+    def dot(
+        self: GenericExpression, other: ndarray
+    ) -> GenericExpression | QuadraticExpression:
         """
         Matrix multiplication with other, similar to xarray dot.
         """
         return self.__matmul__(other)
 
     def __getitem__(
-        self, selector: int | tuple[slice, list[int]] | slice
-    ) -> LinearExpression | QuadraticExpression:
+        self: GenericExpression, selector: int | tuple[slice, list[int]] | slice
+    ) -> GenericExpression:
         """
         Get selection from the expression.
         This is a wrapper around the xarray __getitem__ method. It returns a
@@ -734,15 +664,6 @@ class LinearExpression:
     @property
     def loc(self) -> LocIndexer:
         return LocIndexer(self)
-
-    @classmethod  # type: ignore
-    @property
-    def fill_value(cls) -> dict[str, Any]:
-        warn(
-            "The `.fill_value` attribute is deprecated, use linopy.expressions.FILL_VALUE instead.",
-            DeprecationWarning,
-        )
-        return cls._fill_value
 
     @property
     def type(self) -> str:
@@ -830,42 +751,12 @@ class LinearExpression:
         sol = (self.coeffs * vals).sum(TERM_DIM) + self.const
         return sol.rename("solution")
 
-    @classmethod
-    def _sum(
-        cls,
-        expr: LinearExpression | Dataset,
-        dim: DimsLike | None = None,
-    ) -> Dataset:
-        data = _expr_unwrap(expr)
-
-        if isinstance(dim, str):
-            dim = [dim]
-        elif isinstance(dim, EllipsisType):
-            dim = None
-
-        if dim is None:
-            vars = DataArray(data.vars.data.ravel(), dims=TERM_DIM)
-            coeffs = DataArray(data.coeffs.data.ravel(), dims=TERM_DIM)
-            const = data.const.sum()
-            ds = xr.Dataset({"vars": vars, "coeffs": coeffs, "const": const})
-        else:
-            dim = [d for d in dim if d != TERM_DIM]
-            ds = (
-                data[["coeffs", "vars"]]
-                .reset_index(dim, drop=True)
-                .rename({TERM_DIM: STACKED_TERM_DIM})
-                .stack({TERM_DIM: [STACKED_TERM_DIM] + dim}, create_index=False)
-            )
-            ds = assign_multiindex_safe(ds, const=data.const.sum(dim))
-
-        return ds
-
     def sum(
-        self,
+        self: GenericExpression,
         dim: DimsLike | None = None,
         drop_zeros: bool = False,
         **kwargs: Any,
-    ) -> LinearExpression:
+    ) -> GenericExpression:
         """
         Sum the expression over all or a subset of dimensions.
 
@@ -949,171 +840,6 @@ class LinearExpression:
         dim_dict = {dim_name: self.data.sizes[dim_name] for dim_name in dim}
         return self.rolling(dim=dim_dict).sum(keep_attrs=keep_attrs, skipna=skipna)
 
-    @classmethod
-    def from_tuples(
-        cls, *tuples: tuple, model: Model | None = None
-    ) -> LinearExpression:
-        """
-        Create a linear expression by using tuples of coefficients and
-        variables.
-
-        The function internally checks that all variables in the tuples belong to the same
-        reference model.
-
-        Parameters
-        ----------
-        tuples : tuples of (coefficients, variables)
-            Each tuple represents one term in the resulting linear expression,
-            which can possibly span over multiple dimensions:
-
-            * coefficients : int/float/array_like
-                The coefficient(s) in the term, if the coefficients array
-                contains dimensions which do not appear in
-                the variables, the variables are broadcasted.
-            * variables : str/array_like/linopy.Variable
-                The variable(s) going into the term. These may be referenced
-                by name.
-
-        Returns
-        -------
-        linopy.LinearExpression
-
-        Examples
-        --------
-        >>> from linopy import Model
-        >>> import pandas as pd
-        >>> m = Model()
-        >>> x = m.add_variables(pd.Series([0, 0]), 1)
-        >>> y = m.add_variables(4, pd.Series([8, 10]))
-        >>> expr = LinearExpression.from_tuples((10, x), (1, y))
-
-        This is the same as calling ``10*x + y`` but a bit more performant.
-        """
-        exprs = []
-        for t in tuples:
-            if len(t) == 2:
-                # assume first element is coefficient and second is variable
-                c, v = t
-                if not isinstance(v, (variables.Variable, variables.ScalarVariable)):
-                    raise TypeError("Expected variable as second element of tuple.")
-                expr = v.to_linexpr(c)
-                const = None
-                if model is None:
-                    model = expr.model  # TODO: Ensure equality of models
-            elif len(t) == 1:
-                # assume that the element is a constant
-                c, v = None, None
-                (const,) = as_dataarray(t)
-                if model is None:
-                    raise ValueError("Model must be provided when using constants.")
-                expr = LinearExpression(const, model)
-            else:
-                raise ValueError("Expected tuples of length 1 or 2.")
-
-            exprs.append(expr)
-
-        return merge(exprs, cls=cls) if len(exprs) > 1 else exprs[0]
-
-    @classmethod
-    def from_rule(
-        cls,
-        model: Model,
-        rule: Callable,
-        coords: Sequence[Sequence | pd.Index | DataArray] | Mapping | None = None,
-    ) -> LinearExpression:
-        """
-        Create a linear expression from a rule and a set of coordinates.
-
-        This functionality mirrors the assignment of linear expression as done by
-        Pyomo.
-
-
-        Parameters
-        ----------
-        model : linopy.Model
-            Passed to function `rule` as a first argument.
-        rule : callable
-            Function to be called for each combinations in `coords`.
-            The first argument of the function is the underlying `linopy.Model`.
-            The following arguments are given by the coordinates for accessing
-            the variables. The function has to return a
-            `ScalarLinearExpression`. Therefore use the `.at` accessor when
-            indexing variables.
-        coords : coordinate-like
-            Coordinates to processed by `xarray.DataArray`.
-            For each combination of coordinates, the function
-            given by `rule` is called. The order and size of coords has
-            to be same as the argument list followed by `model` in
-            function `rule`.
-
-
-        Returns
-        -------
-        linopy.LinearExpression
-
-        Examples
-        --------
-        >>> from linopy import Model, LinearExpression
-        >>> m = Model()
-        >>> coords = pd.RangeIndex(10), ["a", "b"]
-        >>> x = m.add_variables(0, 100, coords)
-        >>> def bound(m, i, j):
-        ...     if i % 2:
-        ...         return (i - 1) * x.at[i - 1, j]
-        ...     else:
-        ...         return i * x.at[i, j]
-        ...
-        >>> expr = LinearExpression.from_rule(m, bound, coords)
-        >>> con = m.add_constraints(expr <= 10)
-        """
-        if not isinstance(coords, DataArrayCoordinates):
-            coords = DataArray(coords=coords).coords
-
-        # test output type
-        output = rule(model, *[c.values[0] for c in coords.values()])
-        if not isinstance(output, ScalarLinearExpression) and output is not None:
-            msg = f"`rule` has to return ScalarLinearExpression not {type(output)}."
-            raise TypeError(msg)
-
-        combinations = product(*[c.values for c in coords.values()])
-        exprs = []
-        placeholder = ScalarLinearExpression((np.nan,), (-1,), model)
-        exprs = [rule(model, *coord) or placeholder for coord in combinations]
-        return cls._from_scalarexpression_list(exprs, coords, model)
-
-    @classmethod
-    def _from_scalarexpression_list(
-        cls,
-        exprs: list[ScalarLinearExpression],
-        coords: Mapping,
-        model: Model,
-    ) -> LinearExpression:
-        """
-        Create a LinearExpression from a list of lists with different lengths.
-        """
-        shape = list(map(len, coords.values()))
-
-        coeffs = array(tuple(zip_longest(*(e.coeffs for e in exprs), fillvalue=nan)))
-        vars = array(tuple(zip_longest(*(e.vars for e in exprs), fillvalue=-1)))
-
-        nterm = vars.shape[0]
-        coeffs = coeffs.reshape((nterm, *shape))
-        vars = vars.reshape((nterm, *shape))
-
-        coeffdata = DataArray(coeffs, coords, dims=(TERM_DIM, *coords))
-        vardata = DataArray(vars, coords, dims=(TERM_DIM, *coords))
-        ds = Dataset({"coeffs": coeffdata, "vars": vardata}).transpose(..., TERM_DIM)
-
-        return cls(ds, model)
-
-    def to_quadexpr(self) -> QuadraticExpression:
-        """Convert LinearExpression to QuadraticExpression."""
-        vars = self.data.vars.expand_dims(FACTOR_DIM)
-        fill_value = self._fill_value["vars"]
-        vars = xr.concat([vars, xr.full_like(vars, fill_value)], dim=FACTOR_DIM)
-        data = assign_multiindex_safe(self.data, vars=vars)
-        return QuadraticExpression(data, self.model)
-
     def to_constraint(
         self, sign: SignLike, rhs: ConstantLike | VariableLike | ExpressionLike
     ) -> Constraint:
@@ -1139,7 +865,7 @@ class LinearExpression:
         )
         return constraints.Constraint(data, model=self.model)
 
-    def reset_const(self) -> LinearExpression:
+    def reset_const(self: GenericExpression) -> GenericExpression:
         """
         Reset the constant of the linear expression to zero.
         """
@@ -1157,7 +883,7 @@ class LinearExpression:
         return (self.vars == -1).all(helper_dims) & self.const.isnull()
 
     def where(
-        self,
+        self: GenericExpression,
         cond: DataArray,
         other: LinearExpression
         | int
@@ -1165,7 +891,7 @@ class LinearExpression:
         | dict[str, float | int | DataArray]
         | None = None,
         **kwargs: Any,
-    ) -> LinearExpression | QuadraticExpression:
+    ) -> GenericExpression:
         """
         Filter variables based on a condition.
 
@@ -1187,7 +913,7 @@ class LinearExpression:
 
         Returns
         -------
-        linopy.LinearExpression
+        linopy.LinearExpression or linopy.QuadraticExpression
         """
         # Cannot set `other` if drop=True
         _other: dict[str, float] | dict[str, int | float | DataArray] | DataArray | None
@@ -1196,7 +922,7 @@ class LinearExpression:
                 _other = FILL_VALUE
             else:
                 _other = None
-        elif isinstance(other, (int, float, DataArray)):
+        elif isinstance(other, int | float | DataArray):
             _other = {**self._fill_value, "const": other}
         else:
             _other = _expr_unwrap(other)
@@ -1209,14 +935,14 @@ class LinearExpression:
         return self.__class__(self.data.where(cond, other=_other, **kwargs), self.model)
 
     def fillna(
-        self,
+        self: GenericExpression,
         value: int
         | float
         | DataArray
         | Dataset
         | LinearExpression
         | dict[str, float | int | DataArray],
-    ) -> LinearExpression:
+    ) -> GenericExpression:
         """
         Fill missing values with a given value.
 
@@ -1232,15 +958,15 @@ class LinearExpression:
 
         Returns
         -------
-        linopy.LinearExpression
-            A new `linopy.LinearExpression` object with missing values filled with the given value.
+        linopy.LinearExpression or linopy.QuadraticExpression
+            A new object with missing values filled with the given value.
         """
         value = _expr_unwrap(value)
-        if isinstance(value, (DataArray, np.floating, np.integer, int, float)):
+        if isinstance(value, DataArray | np.floating | np.integer | int | float):
             value = {"const": value}
         return self.__class__(self.data.fillna(value), self.model)
 
-    def diff(self, dim: str, n: int = 1) -> LinearExpression:
+    def diff(self: GenericExpression, dim: str, n: int = 1) -> GenericExpression:
         """
         Calculate the n-th order discrete difference along given axis.
 
@@ -1256,7 +982,7 @@ class LinearExpression:
 
         Returns
         -------
-        linopy.LinearExpression
+        linopy.LinearExpression or linopy.QuadraticExpression
         """
         return self - self.shift({dim: n})
 
@@ -1361,7 +1087,7 @@ class LinearExpression:
         """
         return EmptyDeprecationWrapper(not self.size)
 
-    def densify_terms(self) -> LinearExpression:
+    def densify_terms(self: GenericExpression) -> GenericExpression:
         """
         Move all non-zero term entries to the front and cut off all-zero
         entries in the term-axis.
@@ -1392,7 +1118,7 @@ class LinearExpression:
 
         return self.__class__(data.sel({TERM_DIM: slice(0, nterm)}), self.model)
 
-    def sanitize(self) -> LinearExpression:
+    def sanitize(self: GenericExpression) -> GenericExpression:
         """
         Sanitize LinearExpression by ensuring int dtype for variables.
 
@@ -1405,53 +1131,43 @@ class LinearExpression:
 
         return self
 
-    def equals(self, other: LinearExpression) -> bool:
+    def equals(self, other: BaseExpression) -> bool:
         return self.data.equals(_expr_unwrap(other))
 
     def __iter__(self) -> Iterator[Hashable]:
         return self.data.__iter__()
 
-    @property
-    def flat(self) -> pd.DataFrame:
-        """
-        Convert the expression to a pandas DataFrame.
+    @classmethod
+    def _sum(
+        cls,
+        expr: BaseExpression | Dataset,
+        dim: DimsLike | None = None,
+    ) -> Dataset:
+        data = _expr_unwrap(expr)
+        if cls is QuadraticExpression:
+            dim = dim or list(set(data.dims) - set(HELPER_DIMS))
 
-        The resulting DataFrame represents a long table format of the all
-        expressions with non-zero coefficients. It contains the
-        columns `coeffs` and `vars`.
+        if isinstance(dim, str):
+            dim = [dim]
+        elif isinstance(dim, EllipsisType):
+            dim = None
 
-        Returns
-        -------
-        df : pandas.DataFrame
-        """
-        ds = self.data
+        if dim is None:
+            vars = DataArray(data.vars.data.ravel(), dims=TERM_DIM)
+            coeffs = DataArray(data.coeffs.data.ravel(), dims=TERM_DIM)
+            const = data.const.sum()
+            ds = xr.Dataset({"vars": vars, "coeffs": coeffs, "const": const})
+        else:
+            dim = [d for d in dim if d != TERM_DIM]
+            ds = (
+                data[["coeffs", "vars"]]
+                .reset_index(dim, drop=True)
+                .rename({TERM_DIM: STACKED_TERM_DIM})
+                .stack({TERM_DIM: [STACKED_TERM_DIM] + dim}, create_index=False)
+            )
+            ds = assign_multiindex_safe(ds, const=data.const.sum(dim))
 
-        def mask_func(data: pd.DataFrame) -> pd.Series:
-            mask = (data["vars"] != -1) & (data["coeffs"] != 0)
-            return mask
-
-        df = to_dataframe(ds, mask_func=mask_func)
-        df = df.groupby("vars", as_index=False).sum()
-        check_has_nulls(df, name=self.type)
-        return df
-
-    def to_polars(self) -> pl.DataFrame:
-        """
-        Convert the expression to a polars DataFrame.
-
-        The resulting DataFrame represents a long table format of the all
-        non-masked expressions with non-zero coefficients. It contains the
-        columns `coeffs`, `vars`.
-
-        Returns
-        -------
-        df : polars.DataFrame
-        """
-        df = to_polars(self.data)
-        df = filter_nulls_polars(df)
-        df = group_terms_polars(df)
-        check_has_nulls_polars(df, name=self.type)
-        return df
+        return ds
 
     # Wrapped function which would convert variable to dataarray
     assign = exprwrap(assign_multiindex_safe)
@@ -1511,7 +1227,428 @@ class LinearExpression:
     iterate_slices = iterate_slices
 
 
-class QuadraticExpression(LinearExpression):
+GenericExpression = TypeVar("GenericExpression", bound=BaseExpression)
+
+
+class LinearExpression(BaseExpression):
+    """
+    A linear expression consisting of terms of coefficients and variables.
+
+    The LinearExpression class is a subclass of xarray.Dataset which allows to
+    apply most xarray functions on it. However most arithmetic operations are
+    overwritten. Like this you can easily expand and modify the linear
+    expression.
+
+    Examples
+    --------
+    >>> from linopy import Model
+    >>> import pandas as pd
+    >>> m = Model()
+    >>> x = m.add_variables(pd.Series([0, 0]), 1, name="x")
+    >>> y = m.add_variables(4, pd.Series([8, 10]), name="y")
+
+    Combining expressions:
+
+    >>> expr = 3 * x
+    >>> type(expr)
+    <class 'linopy.expressions.LinearExpression'>
+
+    >>> other = 4 * y
+    >>> type(expr + other)
+    <class 'linopy.expressions.LinearExpression'>
+
+    Multiplying:
+
+    >>> type(3 * expr)
+    <class 'linopy.expressions.LinearExpression'>
+
+    Summation over dimensions
+
+    >>> type(expr.sum(dim="dim_0"))
+    <class 'linopy.expressions.LinearExpression'>
+    """
+
+    @overload
+    def __add__(
+        self,
+        other: ConstantLike | VariableLike | ScalarLinearExpression | LinearExpression,
+    ) -> LinearExpression: ...
+
+    @overload
+    def __add__(self, other: QuadraticExpression) -> QuadraticExpression: ...
+
+    def __add__(
+        self,
+        other: ConstantLike
+        | VariableLike
+        | ScalarLinearExpression
+        | LinearExpression
+        | QuadraticExpression,
+    ) -> LinearExpression | QuadraticExpression:
+        """
+        Add an expression to others.
+
+        Note: If other is a numpy array or pandas object without axes names,
+        dimension names of self will be filled in other
+        """
+        if isinstance(other, QuadraticExpression):
+            return other.__add__(self)
+
+        try:
+            if np.isscalar(other):
+                return self.assign(const=self.const + other)
+
+            other = as_expression(other, model=self.model, dims=self.coord_dims)
+            return merge([self, other], cls=self.__class__)
+        except TypeError:
+            return NotImplemented
+
+    def __radd__(self, other: ConstantLike) -> LinearExpression:
+        try:
+            return self + other
+        except TypeError:
+            return NotImplemented
+
+    @overload
+    def __sub__(
+        self,
+        other: ConstantLike | VariableLike | ScalarLinearExpression | LinearExpression,
+    ) -> LinearExpression: ...
+
+    @overload
+    def __sub__(self, other: QuadraticExpression) -> QuadraticExpression: ...
+
+    def __sub__(
+        self,
+        other: ConstantLike
+        | VariableLike
+        | ScalarLinearExpression
+        | LinearExpression
+        | QuadraticExpression,
+    ) -> LinearExpression | QuadraticExpression:
+        try:
+            return self.__add__(-other)
+        except TypeError:
+            return NotImplemented
+
+    def __rsub__(self, other: ConstantLike | Variable) -> LinearExpression:
+        try:
+            return (self * -1) + other
+        except TypeError:
+            return NotImplemented
+
+    @overload
+    def __mul__(self, other: ConstantLike) -> LinearExpression: ...
+
+    @overload
+    def __mul__(self, other: VariableLike | ExpressionLike) -> QuadraticExpression: ...
+
+    def __mul__(
+        self,
+        other: SideLike,
+    ) -> LinearExpression | QuadraticExpression:
+        """
+        Multiply the expr by a factor.
+        """
+        if isinstance(other, QuadraticExpression):
+            return other.__rmul__(self)
+
+        try:
+            if isinstance(other, variables.Variable | variables.ScalarVariable):
+                other = other.to_linexpr()
+
+            if isinstance(other, LinearExpression | ScalarLinearExpression):
+                return self._multiply_by_linear_expression(other)
+            else:
+                return self._multiply_by_constant(other)
+        except TypeError:
+            return NotImplemented
+
+    def __pow__(self, other: int) -> QuadraticExpression:
+        """
+        Power of the expression with a coefficient. The only coefficient allowed is 2.
+        """
+        if not other == 2:
+            raise ValueError("Power must be 2.")
+        return self * self  # type: ignore
+
+    def __rmul__(self, other: ConstantLike) -> LinearExpression:
+        """
+        Right-multiply the expr by a factor.
+        """
+        try:
+            return self * other
+        except TypeError:
+            return NotImplemented
+
+    @overload
+    def __matmul__(self, other: ConstantLike) -> LinearExpression: ...
+
+    @overload
+    def __matmul__(
+        self, other: VariableLike | ExpressionLike
+    ) -> QuadraticExpression: ...
+
+    def __matmul__(
+        self, other: ConstantLike | VariableLike | ExpressionLike
+    ) -> LinearExpression | QuadraticExpression:
+        """
+        Matrix multiplication with other, similar to xarray dot.
+        """
+        if not isinstance(other, LinearExpression | variables.Variable):
+            other = as_dataarray(other, coords=self.coords, dims=self.coord_dims)
+
+        common_dims = list(set(self.coord_dims).intersection(other.dims))
+        return (self * other).sum(dim=common_dims)
+
+    @property
+    def flat(self) -> pd.DataFrame:
+        """
+        Convert the expression to a pandas DataFrame.
+
+        The resulting DataFrame represents a long table format of the all
+        expressions with non-zero coefficients. It contains the
+        columns `coeffs` and `vars`.
+
+        Returns
+        -------
+        df : pandas.DataFrame
+        """
+        ds = self.data
+
+        def mask_func(data: pd.DataFrame) -> pd.Series:
+            mask = (data["vars"] != -1) & (data["coeffs"] != 0)
+            return mask
+
+        df = to_dataframe(ds, mask_func=mask_func)
+        df = df.groupby("vars", as_index=False).sum()
+        check_has_nulls(df, name=self.type)
+        return df
+
+    def to_quadexpr(self) -> QuadraticExpression:
+        """Convert LinearExpression to QuadraticExpression."""
+        vars = self.data.vars.expand_dims(FACTOR_DIM)
+        fill_value = self._fill_value["vars"]
+        vars = xr.concat([vars, xr.full_like(vars, fill_value)], dim=FACTOR_DIM)
+        data = self.data.assign(vars=vars)
+        return QuadraticExpression(data, self.model)
+
+    def to_polars(self) -> pl.DataFrame:
+        """
+        Convert the expression to a polars DataFrame.
+
+        The resulting DataFrame represents a long table format of the all
+        non-masked expressions with non-zero coefficients. It contains the
+        columns `coeffs`, `vars`.
+
+        Returns
+        -------
+        df : polars.DataFrame
+        """
+        df = to_polars(self.data)
+        df = filter_nulls_polars(df)
+        df = group_terms_polars(df)
+        check_has_nulls_polars(df, name=self.type)
+        return df
+
+    @classmethod
+    def _from_scalarexpression_list(
+        cls,
+        exprs: list[ScalarLinearExpression],
+        coords: Mapping,
+        model: Model,
+    ) -> LinearExpression:
+        """
+        Create a LinearExpression from a list of lists with different lengths.
+        """
+        shape = list(map(len, coords.values()))
+
+        coeffs = array(tuple(zip_longest(*(e.coeffs for e in exprs), fillvalue=nan)))
+        vars = array(tuple(zip_longest(*(e.vars for e in exprs), fillvalue=-1)))
+
+        nterm = vars.shape[0]
+        coeffs = coeffs.reshape((nterm, *shape))
+        vars = vars.reshape((nterm, *shape))
+
+        coeffdata = DataArray(coeffs, coords, dims=(TERM_DIM, *coords))
+        vardata = DataArray(vars, coords, dims=(TERM_DIM, *coords))
+        ds = Dataset({"coeffs": coeffdata, "vars": vardata}).transpose(..., TERM_DIM)
+
+        return cls(ds, model)
+
+    @classmethod
+    def from_rule(
+        cls,
+        model: Model,
+        rule: Callable,
+        coords: Sequence[Sequence | pd.Index | DataArray] | Mapping | None = None,
+    ) -> LinearExpression:
+        """
+        Create a linear expression from a rule and a set of coordinates.
+
+        This functionality mirrors the assignment of linear expression as done by
+        Pyomo.
+
+
+        Parameters
+        ----------
+        model : linopy.Model
+            Passed to function `rule` as a first argument.
+        rule : callable
+            Function to be called for each combinations in `coords`.
+            The first argument of the function is the underlying `linopy.Model`.
+            The following arguments are given by the coordinates for accessing
+            the variables. The function has to return a
+            `ScalarLinearExpression`. Therefore use the `.at` accessor when
+            indexing variables.
+        coords : coordinate-like
+            Coordinates to processed by `xarray.DataArray`.
+            For each combination of coordinates, the function
+            given by `rule` is called. The order and size of coords has
+            to be same as the argument list followed by `model` in
+            function `rule`.
+
+
+        Returns
+        -------
+        linopy.LinearExpression
+
+        Examples
+        --------
+        >>> from linopy import Model, LinearExpression
+        >>> m = Model()
+        >>> coords = pd.RangeIndex(10), ["a", "b"]
+        >>> x = m.add_variables(0, 100, coords)
+        >>> def bound(m, i, j):
+        ...     if i % 2:
+        ...         return (i - 1) * x.at[i - 1, j]
+        ...     else:
+        ...         return i * x.at[i, j]
+        ...
+        >>> expr = LinearExpression.from_rule(m, bound, coords)
+        >>> con = m.add_constraints(expr <= 10)
+        """
+        if not isinstance(coords, DataArrayCoordinates):
+            coords = DataArray(coords=coords).coords
+
+        # test output type
+        output = rule(model, *[c.values[0] for c in coords.values()])
+        if not isinstance(output, ScalarLinearExpression) and output is not None:
+            msg = f"`rule` has to return ScalarLinearExpression not {type(output)}."
+            raise TypeError(msg)
+
+        combinations = product(*[c.values for c in coords.values()])
+
+        placeholder = ScalarLinearExpression((np.nan,), (-1,), model)
+        exprs = [rule(model, *coord) or placeholder for coord in combinations]
+        return cls._from_scalarexpression_list(exprs, coords, model)
+
+    @classmethod
+    def from_tuples(
+        cls,
+        *tuples: tuple[ConstantLike, str | Variable | ScalarVariable] | ConstantLike,
+        model: Model | None = None,
+    ) -> LinearExpression:
+        """
+        Create a linear expression by using tuples of coefficients and
+        variables.
+
+        The function internally checks that all variables in the tuples belong to the same
+        reference model.
+
+        Parameters
+        ----------
+        tuples : A list of elements. Each element is either:
+            * (coefficients, variables)
+            * constant
+
+            Each (coefficients, variables) tuple represents one term in the resulting linear expression,
+            which can possibly span over multiple dimensions:
+
+            * coefficients : int/float/array_like
+                The coefficient(s) in the term, if the coefficients array
+                contains dimensions which do not appear in
+                the variables, the variables are broadcasted.
+            * variables : str/array_like/linopy.Variable
+                The variable(s) going into the term. These may be referenced
+                by name.
+            * constant: int/float/array_like
+                The constant value to add to the expression
+        model : The linopy.Model. If None this can be inferred from the provided variables
+
+        Returns
+        -------
+        linopy.LinearExpression
+
+        Examples
+        --------
+        >>> from linopy import Model
+        >>> import pandas as pd
+        >>> m = Model()
+        >>> x = m.add_variables(pd.Series([0, 0]), 1)
+        >>> y = m.add_variables(4, pd.Series([8, 10]))
+        >>> expr = LinearExpression.from_tuples((10, x), (1, y), 1)
+
+        This is the same as calling ``10*x + y`` + 1 but a bit more performant.
+        """
+
+        def process_one(
+            t: tuple[ConstantLike, str | Variable | ScalarVariable]
+            | tuple[ConstantLike]
+            | ConstantLike,
+        ) -> LinearExpression:
+            nonlocal model
+
+            if isinstance(t, SUPPORTED_CONSTANT_TYPES):
+                if model is None:
+                    raise ValueError("Model must be provided when using constants.")
+                expr = LinearExpression(t, model)
+                return expr
+
+            if not isinstance(t, tuple):
+                raise ValueError("Expected tuple or constant.")
+
+            if len(t) == 2:
+                # assume first element is coefficient and second is variable
+                c, v = t
+                if isinstance(v, variables.ScalarVariable):
+                    if not isinstance(c, int | float):
+                        raise TypeError(
+                            "Expected int or float as coefficient of scalar variable (first element of tuple)."
+                        )
+                    expr = v.to_linexpr(c)
+                elif isinstance(v, variables.Variable):
+                    if not isinstance(c, SUPPORTED_CONSTANT_TYPES):
+                        raise TypeError(
+                            "Expected constant as coefficient of variable (first element of tuple)."
+                        )
+                    expr = v.to_linexpr(c)
+                else:
+                    raise TypeError("Expected variable as second element of tuple.")
+
+                if model is None:
+                    model = expr.model  # TODO: Ensure equality of models
+                return expr
+
+            if len(t) == 1:
+                warn(
+                    "Passing a single value tuple to LinearExpression.from_tuples is deprecated",
+                    DeprecationWarning,
+                )
+                # assume that the element is a constant
+                const = as_dataarray(t[0])
+                if model is None:
+                    raise ValueError("Model must be provided when using constants.")
+                return LinearExpression(const, model)
+            else:
+                raise ValueError("Expected tuples of length 2")
+
+        exprs = [process_one(t) for t in tuples]
+
+        return merge(exprs, cls=cls) if len(exprs) > 1 else exprs[0]
+
+
+class QuadraticExpression(BaseExpression):
     """
     A quadratic expression consisting of terms of coefficients and variables.
 
@@ -1541,35 +1678,42 @@ class QuadraticExpression(LinearExpression):
         data = xr.Dataset(data.transpose(..., FACTOR_DIM, TERM_DIM))
         self._data = data
 
-    def __mul__(
-        self, other: ConstantLike | VariableLike | ExpressionLike
-    ) -> QuadraticExpression:
+    @property
+    def type(self) -> str:
+        return "QuadraticExpression"
+
+    @property
+    def shape(self) -> tuple[int, ...]:
+        # TODO Implement this
+        raise NotImplementedError(
+            f"{self.__class__.__name__} does not support shape property"
+        )
+
+    def __mul__(self, other: SideLike) -> QuadraticExpression:
         """
         Multiply the expr by a factor.
         """
         if isinstance(
             other,
-            (
-                LinearExpression,
-                QuadraticExpression,
-                variables.Variable,
-                variables.ScalarVariable,
-            ),
+            BaseExpression
+            | ScalarLinearExpression
+            | variables.Variable
+            | variables.ScalarVariable,
         ):
             raise TypeError(
                 "unsupported operand type(s) for *: "
                 f"{type(self)} and {type(other)}. "
                 "Higher order non-linear expressions are not yet supported."
             )
-        return super().__mul__(other)  # type: ignore
+        try:
+            return self._multiply_by_constant(other)
+        except TypeError:
+            return NotImplemented
 
-    @property
-    def type(self) -> str:
-        return "QuadraticExpression"
+    def __rmul__(self, other: SideLike) -> QuadraticExpression:
+        return self * other
 
-    def __add__(
-        self, other: ConstantLike | VariableLike | ExpressionLike
-    ) -> QuadraticExpression:
+    def __add__(self, other: SideLike) -> QuadraticExpression:
         """
         Add an expression to others.
 
@@ -1582,27 +1726,20 @@ class QuadraticExpression(LinearExpression):
 
             other = as_expression(other, model=self.model, dims=self.coord_dims)
 
-            if type(other) is LinearExpression:
+            if isinstance(other, LinearExpression):
                 other = other.to_quadexpr()
-            return merge([self, other], cls=self.__class__)  # type: ignore
+
+            return merge([self, other], cls=self.__class__)
         except TypeError:
             return NotImplemented
 
-    def __radd__(
-        self, other: LinearExpression | int
-    ) -> LinearExpression | QuadraticExpression:
+    def __radd__(self, other: ConstantLike) -> QuadraticExpression:
         """
         Add others to expression.
         """
-        if type(other) is LinearExpression:
-            other = other.to_quadexpr()
-            return other.__add__(self)
-        elif other == 0:
-            return self
-        else:
-            return NotImplemented
+        return self.__add__(other)
 
-    def __sub__(self, other: SideLike | QuadraticExpression) -> QuadraticExpression:
+    def __sub__(self, other: SideLike) -> QuadraticExpression:
         """
         Subtract others from expression.
 
@@ -1616,19 +1753,42 @@ class QuadraticExpression(LinearExpression):
             other = as_expression(other, model=self.model, dims=self.coord_dims)
             if type(other) is LinearExpression:
                 other = other.to_quadexpr()
-            return merge([self, -other], cls=self.__class__)  # type: ignore
+            return merge([self, -other], cls=self.__class__)
         except TypeError:
             return NotImplemented
 
-    def __rsub__(self, other: LinearExpression) -> QuadraticExpression:
+    def __rsub__(self, other: SideLike) -> QuadraticExpression:
         """
         Subtract expression from others.
         """
-        if type(other) is LinearExpression:
-            other = other.to_quadexpr()
-            return other.__sub__(self)
-        else:
+        try:
+            return (self * -1) + other
+        except TypeError:
             return NotImplemented
+
+    def __pow__(self, other: SideLike) -> QuadraticExpression:
+        raise TypeError("Higher order non-linear expressions are not yet supported.")
+
+    def __matmul__(
+        self, other: ConstantLike | VariableLike | ExpressionLike
+    ) -> QuadraticExpression:
+        """
+        Matrix multiplication with other, similar to xarray dot.
+        """
+        if isinstance(
+            other,
+            BaseExpression
+            | ScalarLinearExpression
+            | variables.Variable
+            | variables.ScalarVariable,
+        ):
+            raise TypeError(
+                "Higher order non-linear expressions are not yet supported."
+            )
+
+        other = as_dataarray(other, coords=self.coords, dims=self.coord_dims)
+        common_dims = list(set(self.coord_dims).intersection(other.dims))
+        return (self * other).sum(dim=common_dims)
 
     @property
     def solution(self) -> DataArray:
@@ -1641,16 +1801,6 @@ class QuadraticExpression(LinearExpression):
         vals = self._map_solution()
         sol = (self.coeffs * vals.prod(FACTOR_DIM)).sum(TERM_DIM) + self.const
         return sol.rename("solution")
-
-    @classmethod
-    def _sum(
-        cls,
-        expr: Dataset | LinearExpression | QuadraticExpression,
-        dim: DimsLike | None = None,
-    ) -> Dataset:
-        data = _expr_unwrap(expr)
-        dim = dim or list(set(data.dims) - set(HELPER_DIMS))
-        return LinearExpression._sum(expr, dim)
 
     def to_constraint(self, sign: SignLike, rhs: SideLike) -> NotImplementedType:
         raise NotImplementedError(
@@ -1753,9 +1903,9 @@ def as_expression(
     ValueError
         If object cannot be converted to LinearExpression.
     """
-    if isinstance(obj, LinearExpression):
+    if isinstance(obj, LinearExpression | QuadraticExpression):
         return obj
-    elif isinstance(obj, (variables.Variable, variables.ScalarVariable)):
+    elif isinstance(obj, variables.Variable | variables.ScalarVariable):
         return obj.to_linexpr()
     else:
         try:
@@ -1773,9 +1923,9 @@ def merge(
         LinearExpression | QuadraticExpression | variables.Variable | Dataset
     ],
     dim: str = TERM_DIM,
-    cls: type[LinearExpression | QuadraticExpression] = LinearExpression,
+    cls: type[GenericExpression] = None,  # type: ignore
     **kwargs: Any,
-) -> LinearExpression | QuadraticExpression:
+) -> GenericExpression:
     """
     Merge multiple expression together.
 
@@ -1785,7 +1935,6 @@ def merge(
     the coordinates of the first object as a basis which overrides the
     coordinates of the consecutive objects.
 
-
     Parameters
     ----------
     *exprs : tuple/list
@@ -1793,7 +1942,7 @@ def merge(
     dim : str
         Dimension along which the expressions should be concatenated.
     cls : type
-        Type of the resulting expression.
+        Explicitly set the type of the resulting expression (So that the type checker will know the return type)
     **kwargs
         Additional keyword arguments passed to xarray.concat. Defaults to
         {coords: "minimal", compat: "override"} or, in the special case described
@@ -1801,27 +1950,32 @@ def merge(
 
     Returns
     -------
-    res : linopy.LinearExpression
+    res : linopy.LinearExpression or linopy.QuadraticExpression
     """
-    linopy_types = (variables.Variable, LinearExpression, QuadraticExpression)
-
     if not isinstance(exprs, list) and len(add_exprs):
         warn(
             "Passing a tuple to the merge function is deprecated. Please pass a list of objects to be merged",
             DeprecationWarning,
         )
         exprs = [exprs] + list(add_exprs)  # type: ignore
-    model = exprs[0].model
 
-    if (
-        cls is QuadraticExpression
-        and dim == TERM_DIM
-        and any(type(e) is LinearExpression for e in exprs)
-    ):
+    has_quad_expression = any(type(e) is QuadraticExpression for e in exprs)
+    has_linear_expression = any(type(e) is LinearExpression for e in exprs)
+    if cls is None:
+        cls = QuadraticExpression if has_quad_expression else LinearExpression
+
+    if cls is QuadraticExpression and dim == TERM_DIM and has_linear_expression:
         raise ValueError(
             "Cannot merge linear and quadratic expressions along term dimension."
             "Convert to QuadraticExpression first."
         )
+
+    if has_quad_expression and cls is not QuadraticExpression:
+        raise ValueError("Cannot merge linear expressions to QuadraticExpression")
+
+    linopy_types = (variables.Variable, LinearExpression, QuadraticExpression)
+
+    model = exprs[0].model
 
     if cls in linopy_types and dim in HELPER_DIMS:
         coord_dims = [
@@ -1846,6 +2000,8 @@ def merge(
 
         if override:
             kwargs["join"] = "override"
+        else:
+            kwargs.setdefault("join", "outer")
 
     if dim == TERM_DIM:
         ds = xr.concat([d[["coeffs", "vars"]] for d in data], dim, **kwargs)
@@ -1871,7 +2027,7 @@ class ScalarLinearExpression:
     A scalar linear expression container.
 
     In contrast to the LinearExpression class, a ScalarLinearExpression
-    only contains only one label. Use this class to create a constraint
+    only contains one label. Use this class to create a constraint
     in a rule.
     """
 
@@ -1959,7 +2115,7 @@ class ScalarLinearExpression:
         )
 
     def __mul__(self, other: float | int) -> ScalarLinearExpression:
-        if not isinstance(other, (int, float, np.number)):
+        if not isinstance(other, int | float | np.number):
             raise TypeError(
                 f"unsupported operand type(s) for *: {type(self)} and {type(other)}"
             )
@@ -1972,7 +2128,7 @@ class ScalarLinearExpression:
         return self.__mul__(other)
 
     def __div__(self, other: float | int) -> ScalarLinearExpression:
-        if not isinstance(other, (int, float, np.number)):
+        if not isinstance(other, int | float | np.number):
             raise TypeError(
                 f"unsupported operand type(s) for /: {type(self)} and {type(other)}"
             )
@@ -1982,7 +2138,7 @@ class ScalarLinearExpression:
         return self.__div__(other)
 
     def __le__(self, other: int | float) -> AnonymousScalarConstraint:
-        if not isinstance(other, (int, float, np.number)):
+        if not isinstance(other, int | float | np.number):
             raise TypeError(
                 f"unsupported operand type(s) for <=: {type(self)} and {type(other)}"
             )
@@ -1990,7 +2146,7 @@ class ScalarLinearExpression:
         return constraints.AnonymousScalarConstraint(self, LESS_EQUAL, other)
 
     def __ge__(self, other: int | float) -> AnonymousScalarConstraint:
-        if not isinstance(other, (int, float, np.number)):
+        if not isinstance(other, int | float | np.number):
             raise TypeError(
                 f"unsupported operand type(s) for >=: {type(self)} and {type(other)}"
             )
@@ -1998,7 +2154,7 @@ class ScalarLinearExpression:
         return constraints.AnonymousScalarConstraint(self, GREATER_EQUAL, other)
 
     def __eq__(self, other: int | float) -> AnonymousScalarConstraint:  # type: ignore
-        if not isinstance(other, (int, float, np.number)):
+        if not isinstance(other, int | float | np.number):
             raise TypeError(
                 f"unsupported operand type(s) for ==: {type(self)} and {type(other)}"
             )
