@@ -13,11 +13,12 @@ import os
 import re
 import subprocess as sub
 import sys
+import warnings
 from abc import ABC, abstractmethod
 from collections import namedtuple
-from collections.abc import Generator
+from collections.abc import Callable, Generator
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Callable
+from typing import TYPE_CHECKING, Any, Generic, TypeVar
 
 import numpy as np
 import pandas as pd
@@ -32,7 +33,11 @@ from linopy.constants import (
 )
 
 if TYPE_CHECKING:
+    import gurobipy
+
     from linopy.model import Model
+
+EnvType = TypeVar("EnvType")
 
 QUADRATIC_SOLVERS = [
     "gurobi",
@@ -41,6 +46,17 @@ QUADRATIC_SOLVERS = [
     "highs",
     "scip",
     "mosek",
+    "copt",
+    "mindopt",
+]
+
+# Solvers that don't need a solution file when keep_files=False
+NO_SOLUTION_FILE_SOLVERS = [
+    "xpress",
+    "gurobi",
+    "highs",
+    "mosek",
+    "scip",
     "copt",
     "mindopt",
 ]
@@ -87,19 +103,28 @@ with contextlib.suppress(ModuleNotFoundError):
 
     available_solvers.append("cplex")
 
-with contextlib.suppress(ModuleNotFoundError):
+with contextlib.suppress(ModuleNotFoundError, ImportError):
     import xpress
 
     available_solvers.append("xpress")
+
+    # xpress.Namespaces was added in xpress 9.6
+    try:
+        from xpress import Namespaces as xpress_Namespaces
+    except ImportError:
+
+        class xpress_Namespaces:  # type: ignore[no-redef]
+            ROW = 1
+            COLUMN = 2
+            SET = 3
+
 
 with contextlib.suppress(ModuleNotFoundError):
     import mosek
 
     with contextlib.suppress(mosek.Error):
-        with mosek.Env() as m:
-            t = m.Task()
-            t.optimize()
-            m.checkinall()
+        t = mosek.Task()
+        t.optimize()
 
         available_solvers.append("mosek")
 
@@ -196,7 +221,7 @@ def maybe_adjust_objective_sign(
     return solution
 
 
-class Solver(ABC):
+class Solver(ABC, Generic[EnvType]):
     """
     Abstract base class for solving a given linear problem.
 
@@ -243,7 +268,7 @@ class Solver(ABC):
         log_fn: Path | None = None,
         warmstart_fn: Path | None = None,
         basis_fn: Path | None = None,
-        env: None = None,
+        env: EnvType | None = None,
         explicit_coordinate_names: bool = False,
     ) -> Result:
         """
@@ -263,7 +288,7 @@ class Solver(ABC):
         log_fn: Path | None = None,
         warmstart_fn: Path | None = None,
         basis_fn: Path | None = None,
-        env: None = None,
+        env: EnvType | None = None,
     ) -> Result:
         """
         Abstract method to solve a linear problem from a problem file.
@@ -282,7 +307,7 @@ class Solver(ABC):
         log_fn: Path | None = None,
         warmstart_fn: Path | None = None,
         basis_fn: Path | None = None,
-        env: None = None,
+        env: EnvType | None = None,
         explicit_coordinate_names: bool = False,
     ) -> Result:
         """
@@ -323,7 +348,7 @@ class Solver(ABC):
         return SolverName[self.__class__.__name__]
 
 
-class CBC(Solver):
+class CBC(Solver[None]):
     """
     Solver subclass for the CBC solver.
 
@@ -456,7 +481,12 @@ class CBC(Solver):
         status.legacy_status = first_line
 
         # Use HiGHS to parse the problem file and find the set of variable names, needed to parse solution
+        if "highs" not in available_solvers:
+            raise ModuleNotFoundError(
+                f"highspy is not installed. Please install it to use {self.solver_name.name} solver."
+            )
         h = highspy.Highs()
+        h.silent()
         h.readModel(path_to_string(problem_fn))
         variables = {v.name for v in h.getVariables()}
 
@@ -504,7 +534,7 @@ class CBC(Solver):
         return Result(status, solution, CbcModel(mip_gap, runtime))
 
 
-class GLPK(Solver):
+class GLPK(Solver[None]):
     """
     Solver subclass for the GLPK solver.
 
@@ -674,7 +704,7 @@ class GLPK(Solver):
         return Result(status, solution)
 
 
-class Highs(Solver):
+class Highs(Solver[None]):
     """
     Solver subclass for the Highs solver. Highs must be installed
     for usage. Find the documentation at https://www.maths.ed.ac.uk/hall/HiGHS/.
@@ -752,11 +782,11 @@ class Highs(Solver):
             )
 
         h = model.to_highspy(explicit_coordinate_names=explicit_coordinate_names)
+        self._set_solver_params(h, log_fn)
 
         return self._solve(
             h,
             solution_fn,
-            log_fn,
             warmstart_fn,
             basis_fn,
             model=model,
@@ -801,23 +831,35 @@ class Highs(Solver):
 
         problem_fn_ = path_to_string(problem_fn)
         h = highspy.Highs()
+        self._set_solver_params(h, log_fn)
+
         h.readModel(problem_fn_)
 
         return self._solve(
             h,
             solution_fn,
-            log_fn,
             warmstart_fn,
             basis_fn,
             io_api=read_io_api_from_problem_file(problem_fn),
             sense=read_sense_from_problem_file(problem_fn),
         )
 
+    def _set_solver_params(
+        self,
+        highs_solver: highspy.Highs,
+        log_fn: Path | None = None,
+    ) -> None:
+        if log_fn is not None:
+            self.solver_options["log_file"] = path_to_string(log_fn)
+            logger.info(f"Log file at {self.solver_options['log_file']}")
+
+        for k, v in self.solver_options.items():
+            highs_solver.setOptionValue(k, v)
+
     def _solve(
         self,
         h: highspy.Highs,
         solution_fn: Path | None = None,
-        log_fn: Path | None = None,
         warmstart_fn: Path | None = None,
         basis_fn: Path | None = None,
         model: Model | None = None,
@@ -874,13 +916,6 @@ class Highs(Solver):
             highspy.HighsModelStatus.kUnknown: TerminationCondition.unknown,
         }
 
-        if log_fn is not None:
-            self.solver_options["log_file"] = path_to_string(log_fn)
-            logger.info(f"Log file at {self.solver_options['log_file']}")
-
-        for k, v in self.solver_options.items():
-            h.setOptionValue(k, v)
-
         if warmstart_fn is not None and warmstart_fn.suffix == ".sol":
             h.readSolution(path_to_string(warmstart_fn), 0)
         elif warmstart_fn:
@@ -920,7 +955,7 @@ class Highs(Solver):
         return Result(status, solution, h)
 
 
-class Gurobi(Solver):
+class Gurobi(Solver["gurobipy.Env | dict[str, Any] | None"]):
     """
     Solver subclass for the gurobi solver.
 
@@ -943,7 +978,7 @@ class Gurobi(Solver):
         log_fn: Path | None = None,
         warmstart_fn: Path | None = None,
         basis_fn: Path | None = None,
-        env: None = None,
+        env: gurobipy.Env | dict[str, Any] | None = None,
         explicit_coordinate_names: bool = False,
     ) -> Result:
         """
@@ -963,8 +998,8 @@ class Gurobi(Solver):
             Path to the warmstart file.
         basis_fn : Path, optional
             Path to the basis file.
-        env : None, optional
-            Gurobi environment for the solver
+        env : gurobipy.Env or dict, optional
+            Gurobi environment for the solver, pass env directly or kwargs for creation.
         explicit_coordinate_names : bool, optional
             Transfer variable and constraint names to the solver (default: False)
 
@@ -975,6 +1010,8 @@ class Gurobi(Solver):
         with contextlib.ExitStack() as stack:
             if env is None:
                 env_ = stack.enter_context(gurobipy.Env())
+            elif isinstance(env, dict):
+                env_ = stack.enter_context(gurobipy.Env(params=env))
             else:
                 env_ = env
 
@@ -999,7 +1036,7 @@ class Gurobi(Solver):
         log_fn: Path | None = None,
         warmstart_fn: Path | None = None,
         basis_fn: Path | None = None,
-        env: None = None,
+        env: gurobipy.Env | dict[str, Any] | None = None,
     ) -> Result:
         """
         Solve a linear problem from a problem file using the Gurobi solver.
@@ -1018,8 +1055,8 @@ class Gurobi(Solver):
             Path to the warmstart file.
         basis_fn : Path, optional
             Path to the basis file.
-        env : None, optional
-            Gurobi environment for the solver
+        env : gurobipy.Env or dict, optional
+            Gurobi environment for the solver, pass env directly or kwargs for creation.
 
         Returns
         -------
@@ -1032,6 +1069,8 @@ class Gurobi(Solver):
         with contextlib.ExitStack() as stack:
             if env is None:
                 env_ = stack.enter_context(gurobipy.Env())
+            elif isinstance(env, dict):
+                env_ = stack.enter_context(gurobipy.Env(params=env))
             else:
                 env_ = env
 
@@ -1151,7 +1190,7 @@ class Gurobi(Solver):
         return Result(status, solution, m)
 
 
-class Cplex(Solver):
+class Cplex(Solver[None]):
     """
     Solver subclass for the Cplex solver.
 
@@ -1222,6 +1261,17 @@ class Cplex(Solver):
         CONDITION_MAP = {
             "integer optimal solution": "optimal",
             "integer optimal, tolerance": "optimal",
+            "integer infeasible": "infeasible",
+            "time limit exceeded": "time_limit",
+            "time limit exceeded, no integer solution": "infeasible",
+            "error termination": "error",
+            "error termination, no integer solution": "error",
+            "memory limit exceeded": "internal_solver_error",
+            "memory limit exceeded, no integer solution": "internal_solver_error",
+            "aborted": "user_interrupt",
+            "integer unbounded": "unbounded",
+            "integer infeasible or unbounded": "infeasible_or_unbounded",
+            "Unknown status value": "unknown",
         }
         io_api = read_io_api_from_problem_file(problem_fn)
         sense = read_sense_from_problem_file(problem_fn)
@@ -1296,7 +1346,7 @@ class Cplex(Solver):
         return Result(status, solution, m)
 
 
-class SCIP(Solver):
+class SCIP(Solver[None]):
     """
     Solver subclass for the SCIP solver.
 
@@ -1449,7 +1499,7 @@ class SCIP(Solver):
         return Result(status, solution, m)
 
 
-class Xpress(Solver):
+class Xpress(Solver[None]):
     """
     Solver subclass for the xpress solver.
 
@@ -1543,6 +1593,10 @@ class Xpress(Solver):
 
         m.solve()
 
+        # if the solver is stopped (timelimit for example), postsolve the problem
+        if m.getAttrib("solvestatus") == xpress.solvestatus_stopped:
+            m.postsolve()
+
         if basis_fn is not None:
             try:
                 m.writebasis(path_to_string(basis_fn))
@@ -1551,8 +1605,7 @@ class Xpress(Solver):
 
         if solution_fn is not None:
             try:
-                # TODO: possibly update saving of solution file
-                m.writesol(path_to_string(solution_fn))
+                m.writebinsol(path_to_string(solution_fn))
             except Exception as err:
                 logger.info("Unable to save solution file. Raised error: %s", err)
 
@@ -1564,13 +1617,15 @@ class Xpress(Solver):
         def get_solver_solution() -> Solution:
             objective = m.getObjVal()
 
-            var = [str(v) for v in m.getVariable()]
-
-            sol = pd.Series(m.getSolution(var), index=var, dtype=float)
+            var = m.getnamelist(xpress_Namespaces.COLUMN, 0, m.attributes.cols - 1)
+            sol = pd.Series(m.getSolution(), index=var, dtype=float)
 
             try:
-                dual_ = [str(d) for d in m.getConstraint()]
-                dual = pd.Series(m.getDual(dual_), index=dual_, dtype=float)
+                _dual = m.getDual()
+                constraints = m.getnamelist(
+                    xpress_Namespaces.ROW, 0, m.attributes.rows - 1
+                )
+                dual = pd.Series(_dual, index=constraints, dtype=float)
             except (xpress.SolverError, xpress.ModelError, SystemError):
                 logger.warning("Dual values of MILP couldn't be parsed")
                 dual = pd.Series(dtype=float)
@@ -1586,7 +1641,7 @@ class Xpress(Solver):
 mosek_bas_re = re.compile(r" (XL|XU)\s+([^ \t]+)\s+([^ \t]+)| (LL|UL|BS)\s+([^ \t]+)")
 
 
-class Mosek(Solver):
+class Mosek(Solver[None]):
     """
     Solver subclass for the Mosek solver.
 
@@ -1637,8 +1692,9 @@ class Mosek(Solver):
             Path to the warmstart file.
         basis_fn : Path, optional
             Path to the basis file.
-        env : None, optional
-            Mosek environment for the solver
+        env : None, optional, deprecated
+            Deprecated. This parameter is ignored. MOSEK now uses the global
+            environment automatically. Will be removed in a future version.
         explicit_coordinate_names : bool, optional
             Transfer variable and constraint names to the solver (default: False)
 
@@ -1646,24 +1702,27 @@ class Mosek(Solver):
         -------
         Result
         """
-        with contextlib.ExitStack() as stack:
-            if env is None:
-                env_ = stack.enter_context(mosek.Env())
 
-            with env_.Task() as m:
-                m = model.to_mosek(
-                    m, explicit_coordinate_names=explicit_coordinate_names
-                )
+        if env is not None:
+            warnings.warn(
+                "The 'env' parameter in solve_problem_from_model is deprecated and will be "
+                "removed in a future version. MOSEK now uses the global environment "
+                "automatically, avoiding unnecessary license checkouts.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+        with mosek.Task() as m:
+            m = model.to_mosek(m, explicit_coordinate_names=explicit_coordinate_names)
 
-                return self._solve(
-                    m,
-                    solution_fn=solution_fn,
-                    log_fn=log_fn,
-                    warmstart_fn=warmstart_fn,
-                    basis_fn=basis_fn,
-                    io_api="direct",
-                    sense=model.sense,
-                )
+            return self._solve(
+                m,
+                solution_fn=solution_fn,
+                log_fn=log_fn,
+                warmstart_fn=warmstart_fn,
+                basis_fn=basis_fn,
+                io_api="direct",
+                sense=model.sense,
+            )
 
     def solve_problem_from_file(
         self,
@@ -1690,34 +1749,39 @@ class Mosek(Solver):
             Path to the warmstart file.
         basis_fn : Path, optional
             Path to the basis file.
-        env : None, optional
-            Mosek environment for the solver
+        env : None, optional, deprecated
+            Deprecated. This parameter is ignored. MOSEK now uses the global
+            environment automatically. Will be removed in a future version.
 
         Returns
         -------
         Result
         """
-        with contextlib.ExitStack() as stack:
-            if env is None:
-                env_ = stack.enter_context(mosek.Env())
+        if env is not None:
+            warnings.warn(
+                "The 'env' parameter in solve_problem_from_file is deprecated and will be "
+                "removed in a future version. MOSEK now uses the global environment "
+                "automatically, avoiding unnecessary license checkouts.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+        with mosek.Task() as m:
+            # read sense and io_api from problem file
+            sense = read_sense_from_problem_file(problem_fn)
+            io_api = read_io_api_from_problem_file(problem_fn)
+            # for Mosek solver, the path needs to be a string
+            problem_fn_ = path_to_string(problem_fn)
+            m.readdata(problem_fn_)
 
-            with env_.Task() as m:
-                # read sense and io_api from problem file
-                sense = read_sense_from_problem_file(problem_fn)
-                io_api = read_io_api_from_problem_file(problem_fn)
-                # for Mosek solver, the path needs to be a string
-                problem_fn_ = path_to_string(problem_fn)
-                m.readdata(problem_fn_)
-
-                return self._solve(
-                    m,
-                    solution_fn=solution_fn,
-                    log_fn=log_fn,
-                    warmstart_fn=warmstart_fn,
-                    basis_fn=basis_fn,
-                    io_api=io_api,
-                    sense=sense,
-                )
+            return self._solve(
+                m,
+                solution_fn=solution_fn,
+                log_fn=log_fn,
+                warmstart_fn=warmstart_fn,
+                basis_fn=basis_fn,
+                io_api=io_api,
+                sense=sense,
+            )
 
     def _solve(
         self,
@@ -1916,7 +1980,7 @@ class Mosek(Solver):
         return Result(status, solution)
 
 
-class COPT(Solver):
+class COPT(Solver[None]):
     """
     Solver subclass for the COPT solver.
 
@@ -2031,9 +2095,9 @@ class COPT(Solver):
 
         # TODO: check if this suffices
         condition = m.MipStatus if m.ismip else m.LpStatus
-        termination_condition = CONDITION_MAP.get(condition, condition)
+        termination_condition = CONDITION_MAP.get(condition, str(condition))
         status = Status.from_termination_condition(termination_condition)
-        status.legacy_status = condition
+        status.legacy_status = str(condition)
 
         def get_solver_solution() -> Solution:
             # TODO: check if this suffices
@@ -2057,7 +2121,7 @@ class COPT(Solver):
         return Result(status, solution, m)
 
 
-class MindOpt(Solver):
+class MindOpt(Solver[None]):
     """
     Solver subclass for the MindOpt solver.
 
@@ -2195,14 +2259,13 @@ class MindOpt(Solver):
         solution = self.safe_get_solution(status=status, func=get_solver_solution)
         solution = maybe_adjust_objective_sign(solution, io_api, sense)
 
-        # Dispose of the solver to free up the license
         m.dispose()
         env_.dispose()
 
         return Result(status, solution)
 
 
-class PIPS(Solver):
+class PIPS(Solver[None]):
     """
     Solver subclass for the PIPS solver.
     """
