@@ -4,9 +4,10 @@ import numpy as np
 import pandas as pd
 import pytest
 import xarray as xr
+from xarray.testing import assert_equal
 
 from linopy import LESS_EQUAL, Model, Variable
-from linopy.testing import assert_linequal
+from linopy.testing import assert_linequal, assert_quadequal
 
 
 class SomeOtherDatatype:
@@ -94,17 +95,18 @@ def test_arithmetric_operations_variable(m: Model) -> None:
     rng = np.random.default_rng()
     data = xr.DataArray(rng.random(x.shape), coords=x.coords)
     other_datatype = SomeOtherDatatype(data.copy())
-    assert_linequal(x + data, x + other_datatype)  # type: ignore
-    assert_linequal(x - data, x - other_datatype)  # type: ignore
-    assert_linequal(x * data, x * other_datatype)  # type: ignore
+    assert_linequal(x + data, x + other_datatype)
+    assert_linequal(x - data, x - other_datatype)
+    assert_linequal(x * data, x * other_datatype)
     assert_linequal(x / data, x / other_datatype)  # type: ignore
     assert_linequal(data * x, other_datatype * x)  # type: ignore
-    assert x.__add__(object()) is NotImplemented  # type: ignore
-    assert x.__sub__(object()) is NotImplemented  # type: ignore
-    assert x.__mul__(object()) is NotImplemented  # type: ignore
+    assert x.__add__(object()) is NotImplemented
+    assert x.__sub__(object()) is NotImplemented
+    assert x.__mul__(object()) is NotImplemented
     assert x.__truediv__(object()) is NotImplemented  # type: ignore
     assert x.__pow__(object()) is NotImplemented  # type: ignore
-    assert x.__pow__(3) is NotImplemented
+    with pytest.raises(ValueError):
+        x.__pow__(3)
 
 
 def test_arithmetric_operations_expr(m: Model) -> None:
@@ -123,6 +125,17 @@ def test_arithmetric_operations_expr(m: Model) -> None:
     assert expr.__truediv__(object()) is NotImplemented
 
 
+def test_arithmetric_operations_vars_and_expr(m: Model) -> None:
+    x = m.variables["x"]
+    x_expr = x * 1.0
+
+    assert_quadequal(x**2, x_expr**2)
+    assert_quadequal(x**2 + x, x + x**2)
+    assert_quadequal(x**2 * 2, x**2 * 2)
+    with pytest.raises(TypeError):
+        _ = x**2 * x
+
+
 def test_arithmetric_operations_con(m: Model) -> None:
     c = m.constraints["c"]
     x = m.variables["x"]
@@ -133,7 +146,8 @@ def test_arithmetric_operations_con(m: Model) -> None:
     assert_linequal(c.lhs - data, c.lhs - other_datatype)
     assert_linequal(c.lhs * data, c.lhs * other_datatype)
     assert_linequal(c.lhs / data, c.lhs / other_datatype)
-    assert_linequal(c.rhs + data, c.rhs + other_datatype)  # type: ignore
-    assert_linequal(c.rhs - data, c.rhs - other_datatype)  # type: ignore
-    assert_linequal(c.rhs * data, c.rhs * other_datatype)  # type: ignore
-    assert_linequal(c.rhs / data, c.rhs / other_datatype)  # type: ignore
+
+    assert_equal(c.rhs + data, c.rhs + other_datatype)
+    assert_equal(c.rhs - data, c.rhs - other_datatype)
+    assert_equal(c.rhs * data, c.rhs * other_datatype)
+    assert_equal(c.rhs / data, c.rhs / other_datatype)
