@@ -8,8 +8,9 @@ Created on Tue Jan 28 09:03:35 2025.
 from pathlib import Path
 
 import pytest
+from test_io import model  # noqa: F401
 
-from linopy import solvers
+from linopy import Model, solvers
 
 free_mps_problem = """NAME               sample_mip
 ROWS
@@ -64,3 +65,58 @@ def test_free_mps_solution_parsing(solver: str, tmp_path: Path) -> None:
 
     assert result.status.is_ok
     assert result.solution.objective == 30.0
+
+
+@pytest.mark.skipif(
+    "gurobi" not in set(solvers.available_solvers), reason="Gurobi is not installed"
+)
+def test_gurobi_environment_with_dict(model: Model, tmp_path: Path) -> None:  # noqa: F811
+    gurobi = solvers.Gurobi()
+
+    mps_file = tmp_path / "problem.mps"
+    mps_file.write_text(free_mps_problem)
+    sol_file = tmp_path / "solution.sol"
+
+    log1_file = tmp_path / "gurobi1.log"
+    result = gurobi.solve_problem(
+        problem_fn=mps_file, solution_fn=sol_file, env={"LogFile": str(log1_file)}
+    )
+
+    assert result.status.is_ok
+    assert log1_file.exists()
+
+    log2_file = tmp_path / "gurobi2.log"
+    gurobi.solve_problem(
+        model=model, solution_fn=sol_file, env={"LogFile": str(log2_file)}
+    )
+    assert result.status.is_ok
+    assert log2_file.exists()
+
+
+@pytest.mark.skipif(
+    "gurobi" not in set(solvers.available_solvers), reason="Gurobi is not installed"
+)
+def test_gurobi_environment_with_gurobi_env(model: Model, tmp_path: Path) -> None:  # noqa: F811
+    import gurobipy as gp
+
+    gurobi = solvers.Gurobi()
+
+    mps_file = tmp_path / "problem.mps"
+    mps_file.write_text(free_mps_problem)
+    sol_file = tmp_path / "solution.sol"
+
+    log1_file = tmp_path / "gurobi1.log"
+
+    with gp.Env(params={"LogFile": str(log1_file)}) as env:
+        result = gurobi.solve_problem(
+            problem_fn=mps_file, solution_fn=sol_file, env=env
+        )
+
+    assert result.status.is_ok
+    assert log1_file.exists()
+
+    log2_file = tmp_path / "gurobi2.log"
+    with gp.Env(params={"LogFile": str(log2_file)}) as env:
+        gurobi.solve_problem(model=model, solution_fn=sol_file, env=env)
+    assert result.status.is_ok
+    assert log2_file.exists()
