@@ -17,6 +17,7 @@ from xarray.testing import assert_equal
 import linopy
 import linopy.variables
 from linopy import Model
+from linopy.testing import assert_linequal
 
 
 @pytest.fixture
@@ -72,9 +73,7 @@ def test_wrong_variable_init(m: Model, x: linopy.Variable) -> None:
 
 
 def test_variable_getter(x: linopy.Variable, z: linopy.Variable) -> None:
-    with pytest.warns(FutureWarning):
-        assert isinstance(x[0], linopy.variables.ScalarVariable)
-        assert isinstance(z[0], linopy.variables.ScalarVariable)
+    assert isinstance(x[0], linopy.variables.Variable)
 
     assert isinstance(x.at[0], linopy.variables.ScalarVariable)
 
@@ -127,12 +126,12 @@ def test_variable_lower_getter(z: linopy.Variable) -> None:
 
 
 def test_variable_upper_setter(z: linopy.Variable) -> None:
-    z.upper = 20  # type: ignore
+    z.upper = 20
     assert z.upper.item() == 20
 
 
 def test_variable_lower_setter(z: linopy.Variable) -> None:
-    z.lower = 8  # type: ignore
+    z.lower = 8
     assert z.lower == 8
 
 
@@ -181,9 +180,6 @@ def test_variable_sum_warn_unknown_kwargs(x: linopy.Variable) -> None:
 
 def test_fill_value() -> None:
     isinstance(linopy.variables.Variable._fill_value, dict)
-
-    with pytest.warns(DeprecationWarning):
-        linopy.variables.Variable.fill_value
 
 
 def test_variable_where(x: linopy.Variable) -> None:
@@ -305,3 +301,39 @@ def test_variable_iterate_slices(x: linopy.Variable) -> None:
     for s in slices:
         assert isinstance(s, linopy.variables.Variable)
         assert s.size <= 2
+
+
+def test_variable_addition(x: linopy.Variable) -> None:
+    expr1 = x + 1
+    assert isinstance(expr1, linopy.expressions.LinearExpression)
+    expr2 = 1 + x
+    assert isinstance(expr2, linopy.expressions.LinearExpression)
+    assert_linequal(expr1, expr2)
+
+    assert x.__radd__(object()) is NotImplemented
+    assert x.__add__(object()) is NotImplemented
+
+
+def test_variable_subtraction(x: linopy.Variable) -> None:
+    expr1 = -x + 1
+    assert isinstance(expr1, linopy.expressions.LinearExpression)
+    expr2 = 1 - x
+    assert isinstance(expr2, linopy.expressions.LinearExpression)
+    assert_linequal(expr1, expr2)
+
+    assert x.__rsub__(object()) is NotImplemented
+    assert x.__sub__(object()) is NotImplemented
+
+
+def test_variable_multiplication(x: linopy.Variable) -> None:
+    expr1 = x * 2
+    assert isinstance(expr1, linopy.expressions.LinearExpression)
+    expr2 = 2 * x
+    assert isinstance(expr2, linopy.expressions.LinearExpression)
+    assert_linequal(expr1, expr2)
+
+    expr3 = x * x
+    assert isinstance(expr3, linopy.expressions.QuadraticExpression)
+
+    assert x.__rmul__(object()) is NotImplemented
+    assert x.__mul__(object()) is NotImplemented
